+++ packages/plugin-react-app/src/setBase.js
@@ -58,18 +58,12 @@ module.exports = (api, { target, webpackConfig }) => {
       .end()
     // WebpackPluginImport
     .plugin('WebpackPluginImport')
-      .use(WebpackPluginImport, [[
-        {
-          libraryName: /@ali\/ice-.*/,
-          stylePath: 'style.js',
-        },
-      ]])
-      .end();
+      .use(WebpackPluginImport);
+
   if (userConfig.queryLoader) {
     webpackConfig.plugin('WebpackPluginQueryLoader')
       .use(WebpackPluginQueryLoader);
   }
-    
   // auto inject style.js of component (webpack-plugin-import) in mode vite
   if (userConfig.vite) {
     modifyUserConfig('vite.plugins', (vitePlugins) => {